src/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.0
lxml>=5.2.0
sentence-transformers>=3.0.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
